1.2.0.26

+++ Helper/Theme.php
@@ -81,51 +81,48 @@ class Theme extends \Magento\Framework\App\Helper\AbstractHelper
 		$ds = DIRECTORY_SEPARATOR;
 		
 		if (!$this->_path || !is_dir($this->_path)) {
-			IntegrationException::throwException(
-				'Empty or invalid path set.'
-			);
+			IntegrationException::throwException('Empty or invalid path set.');
 		}
-		
-		$targetDir = $this->getTargetDir();
 
-		if (!is_dir($targetDir)) {
-			if (!$this->canAutoInstallTheme()) {
-				IntegrationException::throwException(
-					'The FishPig WordPress theme is not installed. <a href="?install-theme=1">Click here to install it</a>.'
-				);
-			}
-			
-			@mkdir($targetDir, 0777, true);
-			
-			if (!is_dir($targetDir)) {
-				IntegrationException::throwException(
-					'The FishPig WordPress theme is not installed and due to the permissions of the WordPress theme folder, it cannot be installed automatically. Please copy the contents of app/code/FishPig/WordPress/wptheme to the wp-content/themes/fishpig folder.'
-				);
-			}
-		}
-		
+		$targetDir = $this->getTargetDir();
 		$sourceDir = $this->moduleDirReader->getModuleDir('', 'FishPig_WordPress') . $ds . 'wptheme';
-		$sourceFiles = scandir($sourceDir);
 		
-		foreach($sourceFiles as $sourceFile) {
-			if (trim($sourceFile, '.') === '') {
-				continue;
+		$sourceCssFile = $sourceDir . $ds . 'style.css';
+		$targetCssFile = $targetDir . $ds . 'style.css';
+
+		if (!is_dir($targetDir) || !is_file($targetCssFile) || md5_file($sourceCssFile) !== md5_file($targetCssFile)) {
+			// Either theme not installed or version changes
+			if (!is_dir($targetDir)) {
+				@mkdir($targetDir, 0777, true);
+				
+				if (!is_dir($targetDir)) {
+					IntegrationException::throwException(
+						'The FishPig WordPress theme is not installed and due to the permissions of the WordPress theme folder, it cannot be installed automatically. Please copy the contents of app/code/FishPig/WordPress/wptheme to the wp-content/themes/fishpig folder.'
+					);
+				}
 			}
 			
-			$targetFile = $targetDir . self::DS . $sourceFile;
-			$sourceFile = $sourceDir . self::DS . $sourceFile;
+			// Get source files. Loop through and copy to WordPress
+			$sourceFiles = scandir($sourceDir);
 			
-			if (!$this->isFileWriteable($targetFile)) {
-				IntegrationException::throwException(
-					'Unable to install a WordPress theme file due to permissions. File is ' . $targetFile
-				);
-			}
-			
-			$sourceData = file_get_contents($sourceFile);
-			$targetData = file_exists($targetFile) ? file_get_contents($targetFile) : '';
-			
-			if ($sourceData !== $targetData) {
-				file_put_contents($targetFile, $sourceData);
+			foreach($sourceFiles as $sourceFile) {
+				if (trim($sourceFile, '.') === '') {
+					continue;
+				}
+				
+				$targetFile = $targetDir . self::DS . $sourceFile;
+				$sourceFile = $sourceDir . self::DS . $sourceFile;
+				
+				if (!$this->isFileWriteable($targetFile)) {
+					IntegrationException::throwException('Unable to install a WordPress theme file due to permissions. File is ' . $targetFile);
+				}
+				
+				$sourceData = file_get_contents($sourceFile);
+				$targetData = file_exists($targetFile) ? file_get_contents($targetFile) : '';
+				
+				if ($sourceData !== $targetData) {
+					file_put_contents($targetFile, $sourceData);
+				}
 			}
 		}
